Failed pdflatex builds does not ruin script
added '-halt-on-error' flag on default pdflatex builds so that the script can
ignore any commits where the tex file fails to build. Changed check_output call
to 'call' so that no exception is raised. Although in future versions it would
be nice to handle the exception properly and maybe even make it so that the
commit in question is ignored in the future.